refactor: isolate branchless IN-list filter
